Update README.md
Project Components:

TransTibial Prosthesis: The prosthesis is designed to provide support and mobility for individuals with transtibial amputations. 

Mobile Platform: The mobile platform is designed to simulate walking conditions and interaction with the surface during prosthesis use. The platform has one degree of freedom (GDL) and moves along the rotational z-axis.

Interaction between Prosthesis and Platform:

The interaction between the prosthesis and the platform occurs when the prosthesis's keel comes into contact with the platform surface. During the platform's movement, Elastic Foundation Force contact forces are generated, providing dynamic feedback to the prosthesis and simulating realistic walking conditions.

Project Objective:

The objective of this project is to develop a transtibial prosthesis with interaction on a mobile platform. This system is intended to be used as a research and development tool to enhance the effectiveness and comfort of transtibial prostheses.